Added comments and some refactoring to new files
