+++ src/app/core/cache/models/normalized-item.model.ts
@@ -5,6 +5,9 @@ import { Item } from '../../shared/item.model';
 import { mapsTo, relationship } from '../builders/build-decorators';
 import { ResourceType } from '../../shared/resource-type';
 
+/**
+ * Normalized model class for a DSpace Item
+ */
 @mapsTo(Item)
 @inheritSerialization(NormalizedDSpaceObject)
 export class NormalizedItem extends NormalizedDSpaceObject {
@@ -53,6 +56,9 @@ export class NormalizedItem extends NormalizedDSpaceObject {
   @relationship(ResourceType.Collection, false)
   owningCollection: string;
 
+  /**
+   * List of Bitstreams that are owned by this Item
+   */
   @autoserialize
   @relationship(ResourceType.Bitstream, true)
   bitstreams: string[];
